Move some helper functions and classes to math_utils.py
